Move QnnDataType() to utils file
