+++ onnxruntime/core/providers/qnn/builder/opbuilder/cast_op_builder.cc
@@ -6,6 +6,7 @@
 
 #include "core/providers/qnn/builder/qnn_model_wrapper.h"
 #include "core/providers/qnn/builder/op_builder_factory.h"
+#include "core/providers/qnn/builder/qnn_utils.h"
 
 #include "base_op_builder.h"
 
@@ -69,9 +70,9 @@ Status CastOpBuilder::ProcessInputs(QnnModelWrapper& qnn_model_wrapper,
   Qnn_DataType_t qnn_data_type = QNN_DATATYPE_UNDEFINED;
   const auto* type_proto = input.node_arg.TypeAsProto();
 
-  ORT_RETURN_IF_ERROR(GetQnnDataType(false,  // Do not try to get the quantized type. HTP cast supports normal types.
-                                     type_proto,
-                                     qnn_data_type));
+  ORT_RETURN_IF_ERROR(utils::GetQnnDataType(false,  // Do not try to get the quantized type. HTP cast supports normal types.
+                                            type_proto,
+                                            qnn_data_type));
 
   QnnTensorWrapper input_tensorwrapper(input_name, tensor_type, qnn_data_type, QNN_QUANTIZE_PARAMS_INIT,
                                        std::move(input_shape), std::move(unpacked_tensor));
@@ -98,9 +99,9 @@ Status CastOpBuilder::ProcessAttributesAndOutputs(QnnModelWrapper& qnn_model_wra
 
   const auto* type_proto = output.node_arg.TypeAsProto();
   Qnn_DataType_t qnn_data_type = QNN_DATATYPE_UNDEFINED;
-  ORT_RETURN_IF_ERROR(GetQnnDataType(false,  // Do not try to get the quantized type. HTP cast supports normal types.
-                                     type_proto,
-                                     qnn_data_type));
+  ORT_RETURN_IF_ERROR(utils::GetQnnDataType(false,  // Do not try to get the quantized type. HTP cast supports normal types.
+                                            type_proto,
+                                            qnn_data_type));
 
   std::vector<uint32_t> output_shape;
   ORT_RETURN_IF_NOT(qnn_model_wrapper.GetOnnxShape(output.node_arg, output_shape),
