+++ onnxruntime/core/providers/qnn/builder/qnn_utils.h
@@ -23,6 +23,11 @@ std::ostream& operator<<(std::ostream& out, const Qnn_Param_t& qnn_param);
 std::ostream& operator<<(std::ostream& out, const Qnn_Tensor_t& tensor);
 std::ostream& operator<<(std::ostream& out, const QnnOpConfigWrapper& op_conf_wrapper);
 
+Status GetQnnDataType(const bool is_quantized_node, const ONNX_NAMESPACE::TypeProto* type_proto,
+                      Qnn_DataType_t& tensor_data_type);
+
+bool OnnxDataTypeToQnnDataType(const int32_t data_type, Qnn_DataType_t& qnn_data_type, bool is_quantized = false);
+
 }  // namespace utils
 }  // namespace qnn
 }  // namespace onnxruntime
